Add connexion and users permissions

+++ templates/news/show.html.php
@@ -25,7 +25,7 @@
         </div>
     </div>
     <div class="col-lg-3">
-     <div class="col-lg-4">
+       <div class="col-lg-4">
         <div class="card" style="width: 18rem;">
           <div class="card-header">
             A propos de la news
@@ -50,28 +50,34 @@
 <?php } elseif(count($commentaires) === 1){ ?>
     <div class="alert alert-primary" role="alert">Il y a déjà <?= count($commentaires) ?> commentaire. </div>
 <?php } else { ?>
-        <div class="alert alert-primary" role="alert">Il y a déjà <?= count($commentaires) ?> commentaires. </div>
-    <?php } ?>
-    <?php foreach ($commentaires as $commentaire): ?>
-        <div class="container">
-            <div class="row">
-                <div class="col-lg-3" style="border-right: 1px solid <?= Color::rang_etat($commentaire['grade']) ?>">
-                    <div class="avatar-news" style="box-shadow: 0px 0px 2px 2px <?= Color::rang_etat($commentaire['grade']) ?>; background:url('https://www.mangasfan.fr/membres/images/avatars/<?= $commentaire['avatar'] ?>');background-size:100px; background-position: center;"/>
-                    </div>
-                    <p class="pseudo">
-                        <a href="#" style="color: <?= Color::rang_etat($commentaire['grade']); ?>"><?= $commentaire['username']; ?></a><br/>
-                        <span class="badge badge-secondary" style="background-color: <?= Color::rang_etat($commentaire['grade']) ?>;"><?= Color::getRang($commentaire['grade'], $commentaire['chef']) ?></span><br/><br/>
-                        <a href="../news/edit_comment.php?id=<?= $commentaire['id_commentary'] ?>"class="btn btn-sm btn-outline-info">Editer</a>
-                        <a href="../news/delete_comment.php?id=<?= $commentaire['id_commentary'] ?>&news=<?= $news['id_news'] ?>" onclick="return window.confirm(`Êtes vous sûr de vouloir supprimer ce commentaire ?!`)" class="btn btn-sm btn-outline-danger">Supprimer</a>
-                    </p>
+    <div class="alert alert-primary" role="alert">Il y a déjà <?= count($commentaires) ?> commentaires. </div>
+<?php } ?>
+<?php foreach ($commentaires as $commentaire): ?>
+    <div class="container">
+        <div class="row">
+            <div class="col-lg-3" style="border-right: 1px solid <?= Color::rang_etat($commentaire['grade']) ?>">
+                <div class="avatar-news" style="box-shadow: 0px 0px 2px 2px <?= Color::rang_etat($commentaire['grade']) ?>; background:url('https://www.mangasfan.fr/membres/images/avatars/<?= $commentaire['avatar'] ?>');background-size:100px; background-position: center;"/>
                 </div>
-                <div class="col-lg-9">
-                    <?= nl2br($commentaire['commentary']) ?>
-                    <div class="bottom">
-                        <small>Commentaire posté le <?= date('d/m/Y à H:i', strtotime($commentaire['posted_date'])) ?></small>
-                    </div>
-                </div>
-            </div>
+                <p class="pseudo">
+                    <a href="#" style="color: <?= Color::rang_etat($commentaire['grade']); ?>"><?= $commentaire['username']; ?></a><br/>
+                    <span class="badge badge-secondary" style="background-color: <?= Color::rang_etat($commentaire['grade']) ?>;"><?= Color::getRang($commentaire['grade'], $commentaire['chef']) ?></span><br/><br/>
+                    <?php if (isset($_SESSION['auth'])) { 
+                        if ($commentaire['auteur_commentaire'] == $utilisateur['id_user']) {  ?>
+                         <a href="../news/edit_comment.php?id=<?= $commentaire['id_commentary'] ?>"class="btn btn-sm btn-outline-info">Editer</a>
+                         <a href="../news/delete_comment.php?id=<?= $commentaire['id_commentary'] ?>&news=<?= $news['id_news'] ?>" onclick="return window.confirm(`Êtes vous sûr de vouloir supprimer ce commentaire ?!`)" class="btn btn-sm btn-outline-danger">Supprimer</a>
+                     <?php } elseif ($utilisateur['grade'] >= 6 && $utilisateur['grade'] <= 10) { ?>
+                      <a href="../news/delete_comment.php?id=<?= $commentaire['id_commentary'] ?>&news=<?= $news['id_news'] ?>" onclick="return window.confirm(`Êtes vous sûr de vouloir supprimer ce commentaire ?!`)" class="btn btn-sm btn-outline-danger">Supprimer</a>
+                  <?php }
+              } ?>
+          </p>
+      </div>
+      <div class="col-lg-9">
+        <?= nl2br($commentaire['commentary']) ?>
+        <div class="bottom">
+            <small>Commentaire posté le <?= date('d/m/Y à H:i', strtotime($commentaire['posted_date'])) ?></small>
         </div>
-        <hr>
-    <?php endforeach ?>
+    </div>
+</div>
+</div>
+<hr>
+<?php endforeach ?>
